requirements.txt
requests>=2.25.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pytz>=2023.3
